backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
celery==5.4.0
redis==5.2.1
httpx==0.28.1
openai==1.61.1
python-multipart==0.0.20
email-validator==2.2.0
flake8==5.0.4
